menu buttons edited, separate ProfileActivity
