+++ app/src/main/res/layout/activity_main.xml
@@ -4,4 +4,50 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"/>
+    tools:context=".MainActivity">
+
+    <Button
+    android:id="@+id/my_books"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginStart="8dp"
+    android:layout_marginLeft="8dp"
+    android:text="My Books"
+    app:layout_constraintStart_toEndOf="@+id/edit_profile"
+    app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+    android:id="@+id/borrowed"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginStart="8dp"
+    android:layout_marginLeft="8dp"
+    android:text="Borrowed"
+    app:layout_constraintStart_toEndOf="@+id/my_books"
+    app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+    android:id="@+id/edit_profile"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginStart="2dp"
+    android:layout_marginLeft="2dp"
+    android:text="Profile"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+    android:id="@+id/requested"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginStart="8dp"
+    android:layout_marginLeft="8dp"
+    android:layout_marginEnd="2dp"
+    android:layout_marginRight="2dp"
+    android:text="Requested"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintHorizontal_bias="0.6"
+    app:layout_constraintStart_toEndOf="@+id/borrowed"
+    app:layout_constraintTop_toTopOf="parent" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
